requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.2
django-auditlog==3.0.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
